requirements.txt
asgiref==3.2.5
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
Django==2.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pip==20.0.2
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pyodbc==4.0.30
pyparsing==2.4.6
pytz==2019.3
pytest==5.4.1
setuptools==46.0.0
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.8
wrapt==1.11.*
django-storages[azure]
